Make known-agile types generate Send/Sync implementations
If a type has a MarshalingBehaviorAttribute with MarshalingType.Aglile, we know it can be marshalled between threads.

This does not include support objects that implement IAgileObject, which we cannot know at compile time.

The change also includes a basic implementation of the ElementType enum in the interest of sanity, as well as a lot of TypeSig plumbing in service of getting underlying enum types.