gen/function: Emit consistent stub signatures on unsupported platforms
Callers of these functions - whether they're used on unsupported
platforms or not, like `create_instance` with `CoCreateInstance` - fail
to compile when the function signature does not use the new return
pattern for `QueryInterface`-like generic functions.  This change makes
sure the function signature stays the same, only the implementation
changes to `unimplemented!` when the it is not supported on a given
platform.

Fixes: fe8955c ("Transform QueryInterface-like functions into generic functions (#735)")