Find winmd files without reading `PATH` nor copying to `target/`
Non-Windows platforms - which are supported for cross-compiling - do not
set the output directory in `PATH` nor use semicolons to separate this
variable, resulting in errors.
Replace it with the old `include_bytes!` logic that does not require
paths at runtime (of the macro) at all.
Copying of DLL targets to the `target/<profile>` dir for easy
consumption and running of crates is still performed.