gen/win32: Put struct result-pointers in x64 ABI as first parameter
As per [1] the MSVC x64 ABI notes that larger returned datastructures
are allocated by the caller and passed with a pointer.  This is already
implemented in windows-rs, but the return-pointer was always passed as
last argument and the tests never validated a case with more than zero
parameters.  Now that those tests are extended to include extra
parameters and segfault on the functions that return a struct, the
`return__` pointer can be moved as first argument (after `this`) and is
validated to resolve the issue without breaking other tests.

[1]: https://docs.microsoft.com/en-us/cpp/build/x64-calling-convention?view=msvc-160#return-values