Implement `type_name()` for `TypeRef`
When running `tool_windows` on the latest `.winmd` generated with
https://github.com/microsoft/win32metadata/pull/1591 to test the results
of that metadata change, no implementation exists for `TypeRef`.  This
implementation is rather trivial/straightforward based on `TypeDef`.

Going forward, also print the failed value when a `match` statement ends
up in `_ => unimplemented!()` so that it is quicker to immediately debug
what enum variant needs to be implemented.