Use relative paths instead of deriving from CARGO_MANIFEST_DIR.
In the Firefox repository, we avoid vendoring windows-rs because it's
big, and instead pull it some other way. Because cargo doesn't support
this kind of setup, we have a fake windows-rs crate that uses the
modules from where the real windows-rs is. This falls short when the
WindowsAndMessaging feature is enabled because CARGO_MANIFEST_DIR
doesn't point to the real windows-rs.