Split ServiceBuilder off of Service
Service behaves like a singleton, but for the sake of the user, the run()
closure was designed to allow capturing and referencing its outer scope.
This is problematic when trying to pass off the &Service argument to a
thread as the borrow checker complains that the argument escapes the
closure and may not live long enough.

With this change, the argument is now &'static lifetime allowing it to
be freely passed around.